Fix crashing when coming in on a toast notification, and try to repair corrupted profiles
The crash on toast-tap seems to have been caused by the Android NavigationPageRenderer trying to render a toolbar item too early. Moving our intiial navigation out of the constructor and into OnStart seems to have delayed that initial render by long enough for everything to be in place.
Also, if someone only has one profile, if it's missing, we try to repair it on startup, so there's no longer an instacrash.